# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "batsim",
    "description": "Residential battery fleet simulator: physics-faithful virtual homes behind an OpenAPI-first HTTP API for dispatch-strategy testing (ERCOT).",
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/scenarios": {
      "get": {
        "tags": [
          "scenarios"
        ],
        "summary": "List scenarios.",
        "operationId": "list_scenarios",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Page size (1..=1000, default 100).",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 1000,
              "minimum": 1
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque continuation cursor from a previous page.",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(?:[A-Za-z0-9_-]{4})*(?:[A-Za-z0-9_-]{2,3})?$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Page of scenarios",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenariosPage"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "scenarios"
        ],
        "summary": "Create a scenario.",
        "operationId": "create_scenario",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Scenario created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioDoc"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Unavailable data source",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/scenarios/{id}": {
      "get": {
        "tags": [
          "scenarios"
        ],
        "summary": "Get one scenario.",
        "operationId": "get_scenario",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Scenario id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The scenario",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioDoc"
                }
              }
            }
          },
          "404": {
            "description": "Unknown scenario",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/scenarios/{id}:activate": {
      "post": {
        "tags": [
          "scenarios"
        ],
        "summary": "Activate a scenario (requires a stopped simulation).",
        "operationId": "activate_scenario",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Scenario id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Activated scenario",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioDoc"
                }
              }
            }
          },
          "404": {
            "description": "Unknown scenario",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Simulation is not stopped",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Unavailable data source",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/scenarios/{id}:deactivate": {
      "post": {
        "tags": [
          "scenarios"
        ],
        "summary": "Deactivate the active scenario.",
        "operationId": "deactivate_scenario",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Scenario id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Scenario deactivated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkDoc"
                }
              }
            }
          },
          "404": {
            "description": "Unknown scenario",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Simulation is not stopped or scenario not active",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AmbientSpec": {
        "oneOf": [
          {
            "type": "object",
            "description": "Constant temperature.",
            "required": [
              "c",
              "kind"
            ],
            "properties": {
              "c": {
                "type": "number",
                "format": "double",
                "description": "Temperature in deg C."
              },
              "kind": {
                "type": "string",
                "enum": [
                  "constant"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "Diurnal sinusoid peaking mid-afternoon.",
            "required": [
              "mean_c",
              "amplitude_c",
              "kind"
            ],
            "properties": {
              "amplitude_c": {
                "type": "number",
                "format": "double",
                "description": "Half peak-to-peak swing in deg C."
              },
              "kind": {
                "type": "string",
                "enum": [
                  "diurnal"
                ]
              },
              "mean_c": {
                "type": "number",
                "format": "double",
                "description": "Daily mean temperature in deg C."
              }
            }
          }
        ],
        "description": "Ambient temperature feed selection."
      },
...
      "OkDoc": {
        "type": "object",
        "description": "Empty success body for action endpoints with nothing to return.",
        "required": [
          "ok"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "description": "Always true."
          }
        }
      },
...
      "OutageSpec": {
        "type": "object",
        "description": "A scheduled outage window. Recorded with the scenario; outage physics\n(grid loss, islanding) is applied once the resilience milestone ships.",
        "required": [
          "start",
          "end"
        ],
        "properties": {
          "end": {
            "type": "string",
            "format": "date-time",
            "description": "Outage end (RFC 3339 UTC)."
          },
          "load_zones": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Affected load zones (empty = everywhere)."
          },
          "probability": {
            "type": "number",
            "format": "double",
            "description": "Per-home probability of being affected."
          },
          "start": {
            "type": "string",
            "format": "date-time",
            "description": "Outage start (RFC 3339 UTC)."
          }
        },
        "additionalProperties": false
      },
      "PageInfo": {
        "type": "object",
        "description": "Cursor pagination envelope.",
        "required": [
          "has_more"
        ],
        "properties": {
          "has_more": {
            "type": "boolean",
            "description": "Whether more results exist."
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ],
            "description": "Opaque cursor for the next page; null when exhausted."
          }
        }
      },
...
      "PriceSourceSpec": {
        "oneOf": [
          {
            "type": "object",
            "description": "A single flat price for the whole run.",
            "required": [
              "price_per_mwh",
              "source"
            ],
            "properties": {
              "price_per_mwh": {
                "type": "number",
                "format": "double",
                "description": "Price in $/MWh applied at every tick."
              },
              "source": {
                "type": "string",
                "enum": [
                  "static"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "A deterministic synthetic profile.",
            "required": [
              "profile",
              "source"
            ],
            "properties": {
              "amplitude_per_mwh": {
                "type": "number",
                "format": "double",
                "description": "Peak-to-mean swing amplitude in $/MWh."
              },
              "base_price_per_mwh": {
                "type": "number",
                "format": "double",
                "description": "Mean price in $/MWh."
              },
              "profile": {
                "$ref": "#/components/schemas/SyntheticProfile",
                "description": "Profile shape."
              },
              "seed": {
                "type": "integer",
                "description": "Seed reserved for stochastic profile extensions.",
                "maximum": 18446744073709551615,
                "minimum": 0
              },
              "source": {
                "type": "string",
                "enum": [
                  "synthetic"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "Historical replay. Not available yet; activating a scenario with\nthis source fails validation until replay data support lands.",
            "required": [
              "source"
            ],
            "properties": {
              "date_range": {
                "type": [
                  "array",
                  "null"
                ],
                "items": false,
                "prefixItems": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "string"
                  }
                ],
                "description": "Inclusive date range (ISO 8601 dates) to replay."
              },
              "market": {
                "type": [
                  "string",
                  "null"
                ],
                "description": "Market (only the real-time market is modeled)."
              },
              "settlement_point": {
                "type": [
                  "string",
                  "null"
                ],
                "description": "Settlement point / hub."
              },
              "source": {
                "type": "string",
                "enum": [
                  "replay"
                ]
              }
            }
          }
        ],
        "description": "Where settlement prices come from for a scenario."
      },
      "Problem": {
        "type": "object",
        "description": "RFC 9457 problem document; the only error body the API emits.",
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ProblemCode",
            "description": "Stable machine-readable code."
          },
          "detail": {
            "type": [
              "string",
              "null"
            ],
            "description": "Specifics of this occurrence."
          },
          "instance": {
            "type": [
              "string",
              "null"
            ],
            "description": "The request path that produced the problem."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status code.",
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "Short human-readable summary."
          },
          "trace_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Request correlation id (echoed as `X-Request-Id`)."
          },
          "type": {
            "type": "string",
            "description": "Problem type URN."
          }
        },
        "additionalProperties": true
      },
      "ProblemCode": {
        "type": "string",
        "description": "Stable machine-readable error codes.",
        "enum": [
          "VALIDATION_ERROR",
          "UNAUTHORIZED",
          "NOT_FOUND",
          "CONFLICT",
          "UNPROCESSABLE",
          "SIM_RUNNING",
          "SIM_NOT_RUNNING",
          "IDEMPOTENCY_KEY_REUSE",
          "INTERNAL"
        ]
      },
...
      "ScenarioDoc": {
        "type": "object",
        "description": "Scenario document: the binding plus identity and lifecycle.",
        "required": [
          "name",
          "time",
          "prices",
          "seed",
          "id",
          "created_at",
          "active"
        ],
        "properties": {
          "active": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "description": "Scenario name."
          },
          "outages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OutageSpec"
            },
            "description": "Scheduled outages."
          },
          "prices": {
            "$ref": "#/components/schemas/PriceSourceSpec",
            "description": "Price binding."
          },
          "seed": {
            "type": "integer",
            "description": "Scenario seed; governs every random draw during the run.",
            "maximum": 18446744073709551615,
            "minimum": 0
          },
          "time": {
            "$ref": "#/components/schemas/ScenarioTime",
            "description": "Time binding."
          },
          "weather": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/WeatherSpec",
                "description": "Weather binding (default: synthetic 25 C constant)."
              }
            ]
          }
        },
        "additionalProperties": false
      },
      "ScenarioRequest": {
        "type": "object",
        "description": "Scenario creation request (and stored document).",
        "required": [
          "name",
          "time",
          "prices",
          "seed"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Scenario name."
          },
          "outages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OutageSpec"
            },
            "description": "Scheduled outages."
          },
          "prices": {
            "$ref": "#/components/schemas/PriceSourceSpec",
            "description": "Price binding."
          },
          "seed": {
            "type": "integer",
            "description": "Scenario seed; governs every random draw during the run.",
            "maximum": 18446744073709551615,
            "minimum": 0
          },
          "time": {
            "$ref": "#/components/schemas/ScenarioTime",
            "description": "Time binding."
          },
          "weather": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/WeatherSpec",
                "description": "Weather binding (default: synthetic 25 C constant)."
              }
            ]
          }
        },
        "additionalProperties": false
      },
      "ScenarioTime": {
        "type": "object",
        "description": "Scenario time binding.",
        "required": [
          "start",
          "end"
        ],
        "properties": {
          "end": {
            "type": "string",
            "format": "date-time",
            "description": "Run end (RFC 3339 UTC)."
          },
          "start": {
            "type": "string",
            "format": "date-time",
            "description": "Run start (RFC 3339 UTC; must be 5-minute aligned)."
          },
          "tick_seconds": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Tick length in seconds (default 1).",
            "minimum": 0
          }
        },
        "additionalProperties": false
      },
      "ScenariosPage": {
        "type": "object",
        "description": "Page of scenarios.",
        "required": [
          "data",
          "page"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScenarioDoc"
            },
            "description": "Scenarios."
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo",
            "description": "Pagination."
          }
        }
      },
...
      "SyntheticProfile": {
        "type": "string",
        "description": "Synthetic price profile shapes.",
        "enum": [
          "flat",
          "summer_peak"
        ]
      },
...
      "WeatherSpec": {
        "oneOf": [
          {
            "type": "object",
            "description": "Deterministic synthetic ambient feed.",
            "required": [
              "ambient",
              "source"
            ],
            "properties": {
              "ambient": {
                "$ref": "#/components/schemas/AmbientSpec",
                "description": "The ambient feed."
              },
              "source": {
                "type": "string",
                "enum": [
                  "synthetic"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "Historical weather replay. Not available yet.",
            "required": [
              "source"
            ],
            "properties": {
              "dataset": {
                "type": [
                  "string",
                  "null"
                ],
                "description": "Dataset name (reserved)."
              },
              "source": {
                "type": "string",
                "enum": [
                  "replay"
                ]
              }
            }
          }
        ],
        "description": "Weather binding."
      }
    }
  },
...
}